BindinAdapterの実装(StepCountFragmentの可読性が上がりました！)

+++ app/src/main/java/com/hamu/walkforcats/viewmodels/StepCountViewmodel.kt
@@ -25,7 +25,7 @@ class StepCountViewmodel(application: Application): AndroidViewModel(application
     val repository : preferenceRepository = preferenceRepository()
 
     //設定画面や猫部屋から戻るたびに前回分がロードされるのを防ぐための判定に用います。
-    // stepCountFragment が最初に作られ歩数をロードした時にfalse　に変え、 navgraph viewmodel が　clearされる時にtureに戻ります。
+    // stepCountFragment が最初に作られ歩数をロードした時にfalse　に変え、activity viewmodel が　clearされる時にtureに戻ります。
     var isFirstinit = true
     //歩数の1日ごと、週間での集計です
     private val _dailyCount = MutableLiveData(0)
@@ -39,12 +39,12 @@ class StepCountViewmodel(application: Application): AndroidViewModel(application
 
     //1日、１週間の目標です。
     private val _weeklyGoal = MutableLiveData(0)
-    val weeklyGoal: LiveData<Int>
-        get() = _weeklyGoal.map{it.toInt()}
+    val weeklyGoal: LiveData<Float>
+        get() = _weeklyGoal.map{it.toFloat()}
 
     private val _dailyGoal = MutableLiveData(0)
-     val dailyGoal: LiveData<Int>
-        get() = _dailyGoal.map{it.toInt()}
+     val dailyGoal: LiveData<Float>
+        get() = _dailyGoal.map{it.toFloat()}
 
 
     //1日単位、週間単位での歩数の達成率です
